chore: refactored audio mixer panel and created audio track component
